requirements.txt
absl-py==2.1.0
aiofiles==23.1.0
altair==5.4.1
anyio==4.6.0
fastapi==0.115.0
gradio==4.14.0
h5py==3.11.0
huggingface-hub==0.25.1
matplotlib==3.9.0
numpy==1.26.4
opencv-python==4.10.0.84
pandas==2.2.2
requests==2.32.3
scikit-learn==1.5.2
scipy==1.10.1
seaborn==0.13.2
tensorflow==2.17.0
torch==2.5.1  
uvicorn==0.30.6
thop  
protobuf==3.20.3  
keras==3.3.3
transforms==0.2.1
jupyter==1.1.1
ipykernel==6.29.5
streamlit==1.39.0  
plyer==2.1.0
mysql-connector-python==8.0.32 
torchvision 
xformers==0.0.28.post3  
simpleaudio==1.0.4
pytest==7.2.0
torchaudio==2.5.1
